+++ src/usuario/usuario.controller.ts
@@ -1,4 +1,4 @@
-import { Controller,Get, Post, Put, Delete, Res, Body, HttpStatus, Param } from '@nestjs/common';
+import { Controller,Get, Post, Put, Delete, Res, Body, HttpStatus, Param, Req } from '@nestjs/common';
 import { UsuarioService } from './usuario.service';
 import { CreateUsuarioDTO } from './dto/usuario.dto';
 import { Usuario } from './interfaces/usuario.interface';
@@ -17,51 +17,81 @@ export class UsuarioController {
     }
 
     @Post('/create')
-    async createUsuario(@Res() res, @Body() createUsuarioDTO:CreateUsuarioDTO):Promise<CreateUsuarioDTO>{
+    async createUsuario(@Req() req, @Res() res, @Body() createUsuarioDTO:CreateUsuarioDTO):Promise<CreateUsuarioDTO>{
         const passwordHash = await this.getHash(createUsuarioDTO.password);
         try {
-           createUsuarioDTO.password=passwordHash;
-           const usuario = await this.usuarioService.createUsuario(createUsuarioDTO);
-           return res.status(HttpStatus.OK).json(usuario);
+           const cabecera = req.headers['autorization'];
+           const usuarioToken = await this.usuarioService.getToken(cabecera);
+           if (cabecera === usuarioToken.token){
+               createUsuarioDTO.password=passwordHash;
+               const usuario = await this.usuarioService.createUsuario(createUsuarioDTO);
+               return res.status(HttpStatus.OK).json(usuario);
+           }else{
+               return res.status(HttpStatus.UNAUTHORIZED).send(HttpStatus.UNAUTHORIZED);
+           }
         } catch (error) {
            return res.status(HttpStatus.NO_CONTENT).send('<h1>Error</h1><p>Ha ocurrido un problema al crear el usuario</p>');
         }
     };
     @Get()
-    async getUsuarios(@Res() res,):Promise<Usuario>{
+    async getUsuarios(@Req() req, @Res() res,):Promise<Usuario>{
         try {
-            const usuarios = await this.usuarioService.getUsuarios();
-            return res.status(HttpStatus.OK).json(usuarios);
+           const cabecera = req.headers['autorization'];
+           const usuarioToken = await this.usuarioService.getToken(cabecera);
+           if (cabecera === usuarioToken.token){
+                const usuarios = await this.usuarioService.getUsuarios();
+                return res.status(HttpStatus.OK).json(usuarios);
+           }else{
+               return res.status(HttpStatus.UNAUTHORIZED).send(HttpStatus.UNAUTHORIZED);
+           }
         } catch (error) {
             return res.status(HttpStatus.NOT_FOUND).send('<h1>Error</h1><p>No se han encontrado ningún usuario</p>');
         }
     }
     @Get('/:usuarioID')
-    async getUsuario(@Res() res,@Param('usuarioID') usuarioID: string):Promise<Usuario>{
+    async getUsuario(@Req() req, @Res() res,@Param('usuarioID') usuarioID: string):Promise<Usuario>{
         try {
-            const usuario = await this.usuarioService.getUsuario(usuarioID);
-            return res.status(HttpStatus.OK).json(usuario);
+            const cabecera = req.headers['autorization'];
+            const usuarioToken = await this.usuarioService.getToken(cabecera);
+            if (cabecera === usuarioToken.token){
+                const usuario = await this.usuarioService.getUsuario(usuarioID);
+                return res.status(HttpStatus.OK).json(usuario);
+            }else{
+                return res.status(HttpStatus.UNAUTHORIZED).send(HttpStatus.UNAUTHORIZED);
+            }
         } catch (error) {
             return res.status(HttpStatus.NOT_FOUND).send('<h1>Error</h1><p>No se han encontrado ningún usuario</p>');
         }
     }
     @Delete('/delete/:usuarioID')
-    async deleteUsuario(@Res() res,@Param('usuarioID') usuarioID: string){
+    async deleteUsuario(@Req() req, @Res() res,@Param('usuarioID') usuarioID: string){
         try {
-            const usuario = await this.usuarioService.deleteUsuario(usuarioID);
+            const cabecera = req.headers['autorization'];
+            const usuarioToken = await this.usuarioService.getToken(cabecera);
+            if (cabecera === usuarioToken.token){
+                const usuario = await this.usuarioService.deleteUsuario(usuarioID);
             return res.status(HttpStatus.OK).json({
                 mensaje:'Usuario eliminado',
                 usuario
             })
+            }else{
+                return res.status(HttpStatus.UNAUTHORIZED).send(HttpStatus.UNAUTHORIZED);
+            }
         } catch (error) {
             return res.status(HttpStatus.NOT_FOUND).send('<h1>Error</h1><p>Ha ocurrido un problema al intentar eliminar el usuario</p>');
         }
     }
     @Put('/update/:usuarioID')
-    async updateUsuario(@Res() res,@Param('usuarioID') usuarioID: string,@Body() createUsuarioDTO:CreateUsuarioDTO){
+    async updateUsuario(@Req() req, @Res() res,@Param('usuarioID') usuarioID: string,@Body() createUsuarioDTO:CreateUsuarioDTO){
         try {
-            const usuario = await this.usuarioService.updateUsuario(usuarioID,createUsuarioDTO);
-            return res.status(HttpStatus.OK).json(usuario);
+            const cabecera = req.headers['autorization'];
+            const usuarioToken = await this.usuarioService.getToken(cabecera);
+            if (cabecera === usuarioToken.token){
+                const usuario = await this.usuarioService.updateUsuario(usuarioID,createUsuarioDTO);
+                return res.status(HttpStatus.OK).json(usuario);
+            }else{
+                return res.status(HttpStatus.UNAUTHORIZED).send(HttpStatus.UNAUTHORIZED);
+            }
         } catch (error) {
             return res.status(HttpStatus.NOT_FOUND).send('<h1>Error</h1><p>Ha ocurrido un problema al intentar actualizar el usuario</p>')
         }
